move pci to pkg/pci
