+++ pkg/pci/testdata/lspci-vvv
